Resolve conflict in conf.yaml (keep tv_monitors config from dev1.0,
adopt motion_detect fields from main). Fix gocv float32/float64 type
mismatch in detector.go introduced by gocv v0.43.0 upgrade.

+++ internal/tvmonitor/detector.go
@@ -71,7 +71,7 @@ func (d *Detector) TVState(frame gocv.Mat) bool {
 	defer meanDev.Close()
 	gocv.MeanStdDev(diff, &meanDev, &stdDev)
 
-	frameDiff := stdDev.GetFloatAt(0, 0)
+	frameDiff := float64(stdDev.GetFloatAt(0, 0))
 	tvOn = tvOn && frameDiff > d.config.FrameDiffThreshold
 
 	d.prevGray.Close()
